+++ product_code/lagom/authentication_service/CHANGELOG.md
@@ -1,6 +1,7 @@
 # [v.0.5.1 WIP](https://github.com/upb-uc4/University-Credits-4.0/compare/v0.5.0...authentication-v0.5.1) (2020-XX-XX)
 ## Feature
  - Added predefined standard exceptions to CustomException
+ - Added custom deserialization exceptions 
 ## Refactor
  - Changed exceptions to use these standard exceptions whereever possible
 ## Bugfix